requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bqplot==0.12.43
branca==0.7.2
Brotli==1.1.0
build==1.2.1
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.6
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
duckdb==0.10.2
einops==0.8.0
email_validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
Faker==25.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.14.0
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.2.0
geojson==3.1.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.3
ipyleaflet==0.19.1
ipython==8.12.0
ipytree==0.2.2
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-leaflet==0.19.1
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.57
leafmap==0.32.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.2.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.30.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.1
plotly==5.22.0
posthog==3.5.0
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyecharts==2.0.5
Pygments==2.18.0
pymdown-extensions==10.8.1
# pyobjc-core==10.2
# pyobjc-framework-Cocoa==10.2
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pyshp==2.3.1
PySocks==1.7.1
pystac==1.10.1
pystac-client==0.7.7
python-box==7.1.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytrials==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
scooby==0.10.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==2.7.0
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
st-annotated-text==4.0.1
st-copy-to-clipboard==0.1.6
stack-data==0.6.2
starlette==0.37.2
streamlit==1.34.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tenacity==8.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.6.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traittypes==0.2.1
transformers==4.40.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing_extensions==4.11.0
typing-inspect==0.9.0
typing-utils==0.1.0
tzdata==2024.1
ujson==5.9.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
validators==0.28.1
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.17.0
